generator_agent/requirements.txt
langchain_core
langchain-openai
langgraph
langchain_community
fastapi
uvicorn
pydantic
python-dotenv
uvicorn[standard]
httpx